Added files and changed snIa.py to have new functionality
New file:
setup_eups_packages.sh: used to declare and setup packages that are cloned in ${HOME}/soft
Modifications:
Added attributes x1 and c which are being randomly sampled but are not independent at present.
This is because the same random seed is being used to generate both x1 and c and this has to be
changed by obtaining all the uniform random numbers first.