Added files : snIa_t.py and snIaclass_t.py
snIa_t.py : simple version of snIa.py which demonstrates the use of override_formats for raJ2000, and decJ2000 writiing out more decimals and expressing in exponeital form.
snIaclass_t: This is a dynamical class that does everything without using the instance catalog.
snIaclass.py: Not sure what the new modification is ... but it is not working right now.