Added method to draw the SN params from hostid + hostmu
UniversalRules.py: Added method drawSNParams that takes in the hostid
and hostmu and returns c, x1, x0, t0 as a list. This is useful to
recreate a SN found in the catalog (even though the catalog may print
out the SN parameters themselves).

Checked on a few lines that the values of x1, xc, x0, etc. did not
change by diffing lines (though not the whole file, because order was
not maintained)